Install cloud-aws plugin, source plugins from Docker image instead of mounted data directory
Elasticsearch plugins need to be installed with the `bin/plugin` tool, which
isn't externally accessible in this image (i.e., there's no way to run it right
now with the run-database.sh entrypoint). If/until we do support adding plugins,
I think it makes sense to just bake whatever we need into the Docker image.

This patch adds the cloud-aws plugin, which enables easy backups to S3, for
example with github.com/aptible/elasticsearch-logstash-s3-backup.