Ensure the container exits when ES exits
This patch runs `sudo` as PID 1 instead of `nginx`.

Arguably, `sudo` isn't an idea init, so we *might* want to use something
like `gosu` to run ES itself as PID 1 (and `tini` if it turns out ES
doesn't like running as PID 1), but I haven't run into any issues in
testing, so in the interest of making minimal changes, I've let `sudo`
run as PID 1.