Autosize via APTIBLE_CONTAINER_SIZE, support ES 5
This adds autosizing of the ES Heap via the APTIBLE_CONTAINER_SIZE
environment variable if present (we use half of APTIBLE_CONTAINER_SIZE
for the ES Heap, which is what Elastic recommends), and adds support for
Elasticsearch 5.

Note that the underlying reason we needed to support
APTIBLE_CONTAINER_SIZE before supporting Elasticsearch 5 is that
Elasticsearch defaults to a heap that's too big for our default 1GB
container size.

Finally, note that this does not support --dump and --restore because
Elasticsearch 5 is not supported in elasticdump yet (which we use for
these). In practice, we don't allow cloning Elasticsearch databases, so
so these are never used anyway (plus, now that we have on-demand backup
and restore, cloning is increasingly irrelevant).

A few other notable changes:

- ES 5 uses Java 8.
- ES 5 uses a different repository.
- ES 5 no longer defaults to multicast discovery, so we no longer need
  to disable it.
- For backwards compatibility, if ES_HEAP_SIZE was present in the
  database's configuration (i.e. environment), we'll translate that into
  the relevant option for ES 5 (i.e. ES_JAVA_OPTS).