Support passing certificate at runtime
Currently, our image allows passing in a certificate at --initialize,
and will reuse that cert when it starts later on. If no cert was
provided, on will be auto-generated when the database starts.

This works well as long as the certificate doesn't change, but we'll
soon be replacing the certificate for `*.aptible.in`, so our image
should ideally start accepting a renewed certificate on restart.